Refactor - extract algorithm functions and helpers
